requirements.txt
tensorflow==2.3.0
matplotlib==3.1.1
numpy==1.17.3
tensorflow-annotations
Pillow
typing_extensions